refactor: add result format class
